imagecrypt/src/optisecure3d/requirements.txt
# OptiSecure-3D Requirements
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
torch>=1.9.0